Split composites/__init__.py and corresponding tests to separate modules
