+++ satpy/composites/seviri.py
@@ -1,0 +1,54 @@
+# Copyright (c) 2015-2025 Satpy developers
+#
+# This file is part of satpy.
+#
+# satpy is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# satpy is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# satpy.  If not, see <http://www.gnu.org/licenses/>.
+
+"""SEVIRI specific compositors."""
+
+from __future__ import annotations
+
+import logging
+
+from .core import GenericCompositor, IncompatibleAreas
+
+LOG = logging.getLogger(__name__)
+
+
+class RealisticColors(GenericCompositor):
+    """Create a realistic colours composite for SEVIRI."""
+
+    def __call__(self, projectables, *args, **kwargs):
+        """Generate the composite."""
+        projectables = self.match_data_arrays(projectables)
+        vis06 = projectables[0]
+        vis08 = projectables[1]
+        hrv = projectables[2]
+
+        try:
+            ch3 = 3.0 * hrv - vis06 - vis08
+            ch3.attrs = hrv.attrs
+        except ValueError:
+            raise IncompatibleAreas
+
+        ndvi = (vis08 - vis06) / (vis08 + vis06)
+        ndvi = ndvi.where(ndvi >= 0.0, 0.0)
+
+        ch1 = ndvi * vis06 + (1.0 - ndvi) * vis08
+        ch1.attrs = vis06.attrs
+        ch2 = ndvi * vis08 + (1.0 - ndvi) * vis06
+        ch2.attrs = vis08.attrs
+
+        res = super(RealisticColors, self).__call__((ch1, ch2, ch3),
+                                                    *args, **kwargs)
+        return res
